Separate unit and performance test for store cache
